Sync over tasks
